# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/games:
    get:
      operationId: getDashboardGames
      tags: [dashboard]
      summary: Get organiser's games for dashboard
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardGamesResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/payouts:
    get:
      operationId: getDashboardPayouts
      tags: [dashboard]
      summary: Get organiser's payout stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/payout-details:
    post:
      operationId: savePayoutDetails
      tags: [dashboard]
      summary: Save organiser bank payout details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutDetailsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/payout-details/me:
    get:
      operationId: getPayoutDetails
      tags: [dashboard]
      summary: Get organiser's saved payout details
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutDetails"
        "404":
          description: Not set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ==================== AUTH SCHEMAS ====================
...
    DashboardGame:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        pitchName:
          type: string
        kickoffTime:
          type: string
          format: date-time
        price:
          type: number
        capacity:
          type: integer
        status:
          type: string
          enum: [open, full, cancelled]
        bookedCount:
          type: integer
        isPublic:
          type: boolean
        durationMinutes:
          type: integer
      required:
        - id
        - title
        - pitchName
        - kickoffTime
        - price
        - capacity
        - status
        - bookedCount
        - isPublic
        - durationMinutes
    DashboardGamesResponse:
      type: object
      properties:
        upcoming:
          type: array
          items:
            $ref: "#/components/schemas/DashboardGame"
        past:
          type: array
          items:
            $ref: "#/components/schemas/DashboardGame"
        totalUpcoming:
          type: integer
        totalPast:
          type: integer
      required:
        - upcoming
        - past
        - totalUpcoming
        - totalPast
    WeeklyPayout:
      type: object
      properties:
        weekStart:
          type: string
          format: date-time
        weekEnd:
          type: string
          format: date-time
        paidPlayerCount:
          type: integer
        netPayout:
          type: number
        nextPayoutDate:
          type: string
          format: date-time
      required:
        - weekStart
        - weekEnd
        - paidPlayerCount
        - netPayout
        - nextPayoutDate
    PayoutsResponse:
      type: object
      properties:
        weekly:
          $ref: "#/components/schemas/WeeklyPayout"
        allTimeGross:
          type: number
        upcomingGames:
          type: array
          items:
            $ref: "#/components/schemas/DashboardGame"
        payoutDetails:
          $ref: "#/components/schemas/PayoutDetails"
      required:
        - weekly
        - allTimeGross
        - upcomingGames
    PayoutDetailsBody:
      type: object
      properties:
        accountHolder:
          type: string
        iban:
          type: string
        bankName:
          type: ["string", "null"]
        swift:
          type: ["string", "null"]
      required:
        - accountHolder
        - iban
    PayoutDetails:
      type: object
      properties:
        organiserId:
          type: string
        accountHolder:
          type: string
        iban:
          type: string
        bankName:
          type: ["string", "null"]
        swift:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - organiserId
        - accountHolder
        - iban
        - updatedAt

    # ==================== PITCH SCHEMAS ====================
...